refactor(adapter): 테스트 패키지 경로 정리 및 사용하지 않는 import 제거
- 테스트 파일들의 package 선언을 com.reservation.rest.* 하위로 변경하여 패키지 구조를 통일
  - com.reservation.menu -> com.reservation.rest.menu
  - com.reservation.user.general.sign -> com.reservation.rest.user.general.sign
  - com.reservation.category.tag -> com.reservation.rest.category.tag
  - com.reservation.user.general.self -> com.reservation.rest.user.general.self
  - com.reservation.user.seller.sign -> com.reservation.rest.user.seller.sign
  - com.reservation.category.cuisine -> com.reservation.rest.category.cuisine
  - com.reservation.menu 등 기타 테스트 파일들에서 동일한 패키지 경로 조정
- 사용되지 않는 import 문 제거
  - 테스트 파일에서 더 이상 사용되지 않는 MenuUrl, CategoryUrl, RestaurantUrl 등의 import 삭제
- 테스트 소스의 네임스페이스 정합성 및 가독성 개선
