+++ adapter-module/src/test/kotlin/com/reservation/rest/resign/ResignUserControllerTest.kt
@@ -1,0 +1,96 @@
+package com.reservation.rest.resign
+
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.ninjasquad.springmockk.MockkBean
+import com.reservation.config.restdoc.Body
+import com.reservation.config.restdoc.PathParameter
+import com.reservation.config.restdoc.RestDocuments
+import com.reservation.config.security.TestSecurity
+import com.reservation.fixture.CommonlyUsedArbitraries
+import com.reservation.rest.resign.remove.ResignUserController
+import com.reservation.user.resign.port.input.ResignUserUseCase
+import com.reservation.utilities.generator.uuid.UuidGenerator
+import io.kotest.core.spec.style.FunSpec
+import io.kotest.extensions.spring.SpringExtension
+import io.mockk.every
+import org.hamcrest.Matchers.equalTo
+import org.junit.jupiter.api.extension.ExtendWith
+import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
+import org.springframework.context.annotation.Import
+import org.springframework.http.HttpHeaders
+import org.springframework.restdocs.RestDocumentationExtension
+import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.delete
+import org.springframework.restdocs.payload.JsonFieldType
+import org.springframework.test.context.ActiveProfiles
+import org.springframework.test.web.servlet.MockMvc
+import org.springframework.test.web.servlet.result.MockMvcResultHandlers.print
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
+
+@AutoConfigureRestDocs
+@ActiveProfiles(value = ["test"])
+@Import(value = [TestSecurity::class])
+@WebMvcTest(ResignUserController::class)
+@ExtendWith(RestDocumentationExtension::class)
+class ResignUserControllerTest(
+    private val mockMvc: MockMvc,
+    private val objectMapper: ObjectMapper,
+) : FunSpec() {
+    override fun extensions() = listOf(SpringExtension)
+
+    @MockkBean
+    private lateinit var resignUserUseCase: ResignUserUseCase
+
+    init {
+
+        test("사용자가 탈퇴한다.") {
+
+            every {
+                resignUserUseCase.execute(any())
+            } returns true
+
+            val id = UuidGenerator.generate()
+
+            mockMvc.perform(
+                delete(ResignUrl.RESIGN, id)
+                    .header(
+                        HttpHeaders.AUTHORIZATION,
+                        CommonlyUsedArbitraries.bearerTokenArbitrary.sample(),
+                    ),
+            )
+                .andDo(print())
+                .andExpectAll(
+                    status().isResetContent,
+                    jsonPath("$.result").isBoolean,
+                    jsonPath("$.result", equalTo(true)),
+                )
+                .andDo(
+                    RestDocuments(
+                        identifier = "resign",
+                        documentTags = listOf("resign_user"),
+                        summary = "탈퇴",
+                        pathParameter =
+                            arrayOf(
+                                PathParameter(
+                                    name = "id",
+                                    optional = false,
+                                    description = "사용자 ID",
+                                ),
+                            ),
+                        responseBody =
+                            arrayOf(
+                                Body(
+                                    name = "result",
+                                    jsonType = JsonFieldType.BOOLEAN,
+                                    optional = false,
+                                    description = "결과",
+                                ),
+                            ),
+                    )
+                        .authorizedRequestHeader()
+                        .create(),
+                )
+        }
+    }
+}
